Refactored util functions and the Exporter class to support further changes (refs #20)
