+++ src/Core/Utils/IO.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace KKomelin\TranslatableStringExporter\Core;
+namespace KKomelin\TranslatableStringExporter\Core\Utils;
+
+use KKomelin\TranslatableStringExporter\Core\Utils\JSON;
 
 /**
  * Class IO is responsible for reading from and writing to files.
@@ -31,5 +33,16 @@ class IO
         }
 
         return file_get_contents($path);
+    }
+
+    /**
+     * Read existing translation file for the chosen language.
+     *
+     * @param $language_path
+     * @return array
+     */
+    public static function readTranslationFile($language_path) {
+        $content = self::read($language_path);
+        return JSON::jsonDecode($content);
     }
 }
